Create output placeholders with dtype matching scores
This fixes a Float/Double mismatch in the recall tests caused by removing the dtype conversions in `_check_inputs()` to improve performance.

+++ ranking_metrics_torch/common.py
@@ -36,5 +36,5 @@ def _extract_topk(ks, scores, labels):
 
 def _create_output_placeholder(scores, ks):
     return torch.zeros(
-        scores.shape[0], len(ks), dtype=torch.float64, device=scores.device
+        scores.shape[0], len(ks), dtype=scores.dtype, device=scores.device
     )
